+++ library/android-mvc-test/src/main/java/com/shipdream/lib/android/mvc/view/injection/service/StorageService.java
@@ -1,0 +1,19 @@
+package com.shipdream.lib.android.mvc.view.injection.service;
+
+public interface StorageService {
+    class Storage {
+        private String content;
+
+        public String getContent() {
+            return content;
+        }
+
+        public void setContent(String content) {
+            this.content = content;
+        }
+    }
+
+    String getContent();
+
+    void setContent(String content);
+}