+++ ChangeLog.md
@@ -1,3 +1,7 @@
+Version:1.5.3
+* MvcGraph able to inject concrete class with a public constructor
+* Fix bug that sub fragments' controller do not restore state
+
 Version:1.5.2
 Fix bug to ensure all fragments are associated with correct activity